Abstracted map loading to DefaultMapLoader : IMapLoader
